Refactor combat and strike to use optional and to use lambdas so that skills will be easier in the future.

+++ Project1/IterableBitset.h
@@ -51,7 +51,7 @@ public:
 		return _values;
 	}
 
-	auto operator[](int i) {
+	auto operator[](size_t i) {
 		return _values[i];
 	}
